src/kookin/requirements.txt
mcdreforged>=2.6.0
khl.py>=0.3.17
